refactor: Rename `cli_utils` source and header files
This makes them consistent with the overriding file naming convention.
