+++ lib/cliutils.cc
