+++ plpfuse/main.cc
@@ -23,7 +23,7 @@
 
 #include <bufferarray.h>
 #include <bufferstore.h>
-#include <cli_utils.h>
+#include <cliutils.h>
 #include <cstdint>
 #include <drive.h>
 #include <rfsv.h>
